requirements.txt
mprpc==0.1.16
python-etcd==0.4.5
